+++ python/lsst/ts/genericcamera/__init__.py
@@ -1,4 +1,4 @@
-# This file is part of ts_GenericCamera.
+# This file is part of ts_genericcamera.
 #
 # Developed for the Vera Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
@@ -19,9 +19,15 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .genericcamera import *
-from .andorcamera import *
-from .canoncamera import *
-from .simulatorcamera import *
-from .zwocamera import *
-from .zwofilterwheel import *
+try:
+    from .version import *
+except ModuleNotFoundError:
+    __version__ = "?"
+
+from .config_schema import *
+from .genericcameracsc import *
+from .exposure import *
+from .fits_header_items_generator import *
+from .liveview import *
+from .driver import *
+from .utils import *