+++ python/lsst/ts/genericcamera/driver/__init__.py
@@ -1,4 +1,4 @@
-# This file is part of ts_GenericCamera.
+# This file is part of ts_genericcamera.
 #
 # Developed for the Vera Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
@@ -18,3 +18,10 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+from .genericcamera import *
+from .andorcamera import *
+from .canoncamera import *
+from .simulatorcamera import *
+from .zwocamera import *
+from .zwofilterwheel import *