+++ python/lsst/ts/genericcamera/headers/__init__.py
@@ -1,4 +1,4 @@
-# This file is part of ts_GenericCamera.
+# This file is part of ts_genericcamera.
 #
 # Developed for the Vera Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
@@ -18,16 +18,3 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-try:
-    from .version import *
-except ModuleNotFoundError:
-    __version__ = "?"
-
-from .config_schema import *
-from .genericcameracsc import *
-from .exposure import *
-from .fits_header_items_generator import *
-from .liveview import *
-from .driver import *
-from .utils import *